+++ src/main/java/br/com/phmiranda/oracle/objetos/banco/base/Conta.java
@@ -5,11 +5,8 @@
  * User: phmiranda
  */
 
-package br.com.phmiranda.oracle.objeto.banco.conta;
+package br.com.phmiranda.oracle.objetos.banco.base;
 
-public class SeguroVida implements Tributavel {
-    @Override
-    public double getValorImposto() {
-        return 42;
-    }
+public class Conta {
+
 }
