+++ src/main/java/br/com/phmiranda/oracle/objetos/banco/conta/CalculadoraImposto.java
@@ -5,7 +5,7 @@
  * User: phmiranda
  */
 
-package br.com.phmiranda.oracle.objeto.banco.conta;
+package br.com.phmiranda.oracle.objetos.banco.conta;
 
 public class CalculadoraImposto {
     private double totalImposto;
