+++ src/main/java/br/com/phmiranda/oracle/objetos/funcionario/Cliente.java
@@ -5,7 +5,7 @@
  * User: phmiranda
  */
 
-package br.com.phmiranda.oracle.objeto.funcionario;
+package br.com.phmiranda.oracle.objetos.funcionario;
 
 public class Cliente implements Autenticavel {
     private AutenticacaoUtil autenticador;
